Add explicit int conversion

+++ src/eo_matrix.h
@@ -56,7 +56,7 @@ inline void removeColumn(dMat& matrix, unsigned int colToRemove)
 
 inline void removeElement(dVec &vec, unsigned int k)
 {
-  if (k >= vec.size())
+  if ((int) k >= vec.size())
     return;
   std::copy(vec.data()+k+1,vec.data()+vec.size(),vec.data()+k);
   vec.conservativeResize(vec.size()-1);
@@ -64,7 +64,7 @@ inline void removeElement(dVec &vec, unsigned int k)
 
 inline void addElement(dVec &vec, unsigned int k, const double &xk)
 {
-  if (k >= vec.size()+1)
+  if ((int) k >= vec.size()+1)
     return;
   vec.conservativeResize(vec.size()+1);
   std::copy(vec.data()+k,vec.data()+vec.size()-1,vec.data()+k+1);
